html/semantics/sections/the-body-element/sections-body-element-events-methods.html

<!DOCTYPE html>
<html>
<head>
    <meta charset="utf-8">
    <title>Body Element</title>
    <link rel="author" title="dzenana" href="mailto:dzenana.trenutak@gmail.com">
    <link rel="help" href="http://www.w3.org/html/wg/drafts/html/master/single-page.html#the-body-element">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
<body>

    <h1>Description</h1>
    <p>This test validates that all automable (16 out of 20) body element IDL event handlers react to relevant events.</p>

    <div id="log"></div>

    <script>
        "use strict";

	// Verify that all the content attributes that allow body to react to certain events actually do react to those events

	var myReaction = [], num = 0, strEvent="";
	var myEvents = [["error", true, false],["hashchange", true, false], ["resize", true, false], ["scroll", true, false], ["beforeunload", false, true], ["blur", false, false], ["focus", false, false], ["load", false, false], ["unload", false, false], ["fullscreenchange", false, false], ["fullscreenerror", false, false], ["pagehide", false, false], ["pageshow", false, false], ["popstate", false, false], ["storage", false, false]];

	for (num = 0; num < myEvents.length; num++) {

		strEvent = myEvents[num][0];

		test(function() {

			var myEvent = {};

			// create event handler using document.body.property
			document.body["on"+strEvent] = function() { myReaction.push(strEvent); }; 

			// double-register event with event listener to make sure something happened for content attribute to react to
			document.body.addEventListener(strEvent, function(event) { myReaction.push(strEvent + "-listener"); }, false);

			// Create and cause the event
			myEvent = document.createEvent("Event");
                	myEvent.initEvent(strEvent, myEvents[num][1], myEvents[num][2]);
			document.body.dispatchEvent(myEvent);
 
			// Check that event fired by looking for event listener results
			assert_true(myReaction.indexOf(strEvent + "-listener") > -1, strEvent + " DID NOT FIRE: ");			

			// Check that content attribute's event handler fired
			assert_true(myReaction.indexOf(strEvent) > -1, "on" + strEvent + " IDL method failed: ");
             	
            	}, "Body's on" + strEvent + " IDL method responds to event.");

	}

	// Now handle the message event asynchronously
	var msgTest = async_test("Body's onmessage IDL method responds to event.");

	msgTest.step(function () {
		
		document.body.onmessage = function () { myReaction.push("message"); };

		// add this event listener as a triple-check that event fires and is caught by this test
		window.addEventListener("message", function(event) { 
			myReaction.push("message-WINDOW"); 
			console.log("message-WINDOW:  " + myReaction.toString() + " : " + myReaction.indexOf("message-WINDOW"));
		}, false);
 	});

	msgTest.step(function () {
		window.postMessage("message_here", "*");
    	});

	msgTest.step(function () {
		window.onmessage = msgTest.step_func(function (event) {
			assert_true(myReaction.indexOf("message-WINDOW") > -1, "MESSAGE EVENT NOT REGISTERED BY WINDOW EV LISTENER: ");        	
			assert_true(myReaction.indexOf("message") > -1, "onmessage IDL method failed: ");
			msgTest.done();
		});
    	});

    </script>

</body>
